+++ tomcat/src/test/java/org/apache/coyote/http11/http/ContentTypeTest.java
@@ -1,9 +1,8 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.http;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import org.apache.coyote.http11.http.ContentType;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
